Separate integration tests
